client/pyproject.toml
[tool.poetry]
name = "client"
version = "1.0.0"
description = "Aplicação cliente para transmissão de vídeo e chat."
authors = ["PitzTech <victorlaurentino7@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.scripts]
start = "src.main:main"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.10.0.82"
flask = "^3.0.3"
flask-socketio = "^5.3.6"
python-socketio = {extras = ["client"], version = "^5.3.0"}
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
numpy = "^1.26.4"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
